added rowCount to routes

+++ src/models/api-schema.sql
@@ -5,12 +5,10 @@ CREATE TABLE clothes(
     id SERIAL PRIMARY KEY,
     name VARCHAR(255) NOT NULL,
     description VARCHAR(255),
-    price INTEGER(255)
-);
+    price INTEGER);
 
 CREATE TABLE food(
     id SERIAL PRIMARY KEY,
     name VARCHAR(255) NOT NULL,
     description VARCHAR(255),
-    price INTEGER(255)
-);
+    price INTEGER);
